morning update 5112022

+++ code/python_scripts/labfunction-calculate2.py
@@ -37,8 +37,8 @@ def calculate2(num1, num2, op):  # num1, num2, op
 if len(sys.argv) < 4:
     exit("Usage: " + sys.argv[0] + " OPERAND OPERAND OPERATOR")
 
-num1 = int(sys.argv[1])
-num2 = int(sys.argv[2])
+num1 = float(sys.argv[1])
+num2 = float(sys.argv[2])
 op = sys.argv[3]
 
 answer = calculate2(num1, num2, op)
